Move aesthetic improvements from ODKOptionPane to UI - Provide meaningful reusable factories for the parent Dialog object and the ScrollPane that will hold the actual message
