+++ src/org/opendatakit/briefcase/ui/reused/UI.java
@@ -16,17 +16,25 @@
 
 package org.opendatakit.briefcase.ui.reused;
 
+import static java.awt.Color.GRAY;
 import static java.awt.Color.LIGHT_GRAY;
 import static javax.swing.JOptionPane.ERROR_MESSAGE;
 import static javax.swing.JOptionPane.getFrameForComponent;
 import static org.opendatakit.briefcase.ui.ScrollingStatusListDialog.showDialog;
 
+import java.awt.Dimension;
 import java.awt.Font;
 import java.awt.Insets;
+import java.awt.event.ComponentAdapter;
+import java.awt.event.ComponentEvent;
+import javax.swing.BorderFactory;
 import javax.swing.JButton;
 import javax.swing.JDialog;
+import javax.swing.JLabel;
 import javax.swing.JOptionPane;
+import javax.swing.JScrollPane;
 import javax.swing.JTable;
+import javax.swing.JTextArea;
 import javax.swing.SwingWorker;
 import org.opendatakit.briefcase.model.FormStatus;
 
@@ -61,18 +69,65 @@ public class UI {
   }
 
   public static void errorMessage(String title, String message, boolean nonBlocking) {
-    JDialog dialog = new JDialog();
-    dialog.setAlwaysOnTop(true);
     if (nonBlocking) {
       new SwingWorker() {
         @Override
         protected Object doInBackground() {
-          JOptionPane.showMessageDialog(dialog, message, title, ERROR_MESSAGE);
+          JOptionPane.showMessageDialog(buildDialogParent(), buildScrollPane(message), title, ERROR_MESSAGE);
           return null;
         }
       }.execute();
     } else {
-      JOptionPane.showMessageDialog(dialog, message, title, ERROR_MESSAGE);
+      JOptionPane.showMessageDialog(buildDialogParent(), buildScrollPane(message), title, ERROR_MESSAGE);
     }
+  }
+
+  private static JDialog buildDialogParent() {
+    JDialog dialog = new JDialog();
+    // We want all dialogs show on top
+    dialog.setAlwaysOnTop(true);
+    return dialog;
+  }
+
+  private static JScrollPane buildScrollPane(String message) {
+    // create a n-character wide label for aiding layout calculations...
+    // the dialog box will display this width of text.
+    JLabel probe = new JLabel("MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM");
+
+    JTextArea textArea = new JTextArea(message);
+    textArea.setEditable(false);
+    textArea.setLineWrap(true);
+    textArea.setWrapStyleWord(true);
+    // Take font and colors from the probe
+    textArea.setFont(probe.getFont());
+    textArea.setBackground(probe.getBackground());
+    textArea.setForeground(probe.getForeground());
+
+    JScrollPane scrollPane = new JScrollPane(textArea);
+
+    // don't show the gray border of the scroll pane
+    // unless we are showing the scroll bar, in which case we do show it.
+    scrollPane.setBorder(BorderFactory.createEmptyBorder());
+    scrollPane.getVerticalScrollBar().addComponentListener(new ComponentAdapter() {
+      @Override
+      public void componentHidden(ComponentEvent component) {
+        if (component.getComponent().equals(scrollPane.getVerticalScrollBar())) {
+          scrollPane.setBorder(BorderFactory.createEmptyBorder());
+        }
+      }
+
+      @Override
+      public void componentShown(ComponentEvent component) {
+        if (component.getComponent().equals(scrollPane.getVerticalScrollBar())) {
+          scrollPane.setBorder(BorderFactory.createLineBorder(GRAY));
+        }
+      }
+    });
+
+    Dimension dimension = probe.getPreferredSize();
+    dimension.setSize(dimension.getWidth(), 5.3 * dimension.getHeight());
+    scrollPane.setMinimumSize(dimension);
+    scrollPane.setPreferredSize(dimension);
+    return scrollPane;
   }
 }
